backend/requirements.txt
# Core LLM + Agent framework
langchain==0.3.19
langchain-openai==0.3.7
langgraph==0.3.5

# Web search
tavily-python==0.5.0

# Web framework
fastapi==0.115.11
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
supabase==2.13.0

# Background scheduling (monitor polls)
APScheduler==3.10.4

# Email notifications
resend==2.9.0

# Utilities
python-dotenv==1.0.1
pydantic==2.10.6
httpx==0.28.1
